Add files via upload
1. Login page
I've assumed that there is one admin and his/her username and password is already in the Admin database. beside this the users are assumed to be use this system when only the admin add them to the system.
It which includes login as Admin, Librarian and user. Check the input and go to the required page.
2. Admin page
The admin page contains manage Librarian and manage Users, which add Librarian and Users and manage them.
3. Librarian page
Librarian has a privilege of add books, issue books, return books and so on.
4. User page
The User can issue books and be served.

Beside this the system has different pages that facilitate the service